pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "etl"
version = "0.1.0"
description = "Cookiecutter-style ETL pipeline"
authors = [{ name = "Your Name" }]
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "pandas",
    "sqlalchemy",
    "pyarrow",
    "python-dotenv",
    "requests",
    "tabulate"
]